A lot of changes

+++ BLL/DomainModel/Factors/Medicine/BLOs/MedicineRuleItem.cs
@@ -12,7 +12,7 @@ namespace BLL.DomainModel.Factors.Medicine.BLOs
         public virtual int UnitDoseQuantifier { get; set; }
         public virtual bool HasUserDefinedUnitDose { get; set; }
         public virtual UserDefinedUnitDoseType? UserDefinedUnitDoseType { get; set; }
-        public virtual int? UserDefinedUnitDoseSize { get; set; }
+        public virtual decimal? UserDefinedUnitDoseSize { get; set; }
 
     }
