+++ BLL/DomainModel/Factors/Medicine/BLOs/MedicineType.cs
@@ -14,7 +14,7 @@ namespace BLL.DomainModel.Factors.Medicine.BLOs
         virtual public UnitOfMeasure BaseUnitOfMeasure { get; set; }
         virtual public bool IsPackagedIntoUnits { get; set; }
         virtual public PackagedUnitDoseType? PackagedUnitDoseType { get; set; }
-        virtual public int? PackagedUnitDoseSize { get; set; }
+        virtual public Decimal? PackagedUnitDoseSize { get; set; }
 
         virtual public bool? IsInUse { get; set; }
         virtual public int? CurrentSupplyAmount { get; set; } // Null if no SupplyEntries exist
